requirements.txt
pyinstaller
requests
pillow
numpy
opencv-python
pycryptodome
pywin32
pyperclip
pynput
pyaudio
scapy